fix: Authentication issues
