+++ front/package.json
@@ -21,8 +21,8 @@
     "@aleph-sdk/ethereum": "1.0.3",
     "@aleph-sdk/evm": "^1.2.0",
     "@aleph-sdk/message": "1.0.7",
-    "@privy-io/react-auth": "^1.92.2",
-    "@privy-io/wagmi": "^0.2.12",
+    "@privy-io/react-auth": "^2.2.1",
+    "@privy-io/wagmi": "^1.0.2",
     "@radix-ui/react-avatar": "^1.1.1",
     "@radix-ui/react-checkbox": "^1.1.2",
     "@radix-ui/react-context-menu": "^2.2.2",
